Added Locations
Added locatiosn and location schema

+++ database/migrations/2017_12_28_234510_create_campuses_table.php
@@ -16,8 +16,8 @@ class CreateCampusesTable extends Migration
         Schema::create('campuses', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name')->isNullable(false);
-            $table->decimal('longitude')->isNullable(false);
-            $table->decimal('latitude')->isNullable(false);
+            $table->double('longitude', 15, 8)->isNullable(false);
+            $table->double('latitude', 15, 8)->isNullable(false);
             $table->timestamps();
         });
     }
